# specs/001-core-ventas-inventario/contracts/core-ventas-inventario.openapi.yaml
openapi: 3.0.3
info:
  title: El Kiosquito — Core de Ventas e Inventario
  version: 1.5.0
  description: Contrato de la API del módulo 001-core-ventas-inventario. Ver spec.md para requisitos y data-model.md para el esquema completo.
servers:
  - url: /api/v1
paths:
...
  /productos:
    post:
      summary: Registrar producto (RF-CVI-006, RF-CVI-007)
      operationId: crearProducto
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nombre, categoria_id, unidad_venta_codigo, unidad_inventario_codigo, es_fraccionable]
              properties:
                nombre: { type: string }
                categoria_id: { type: integer, description: 'FK a categoria — RF-CVI-023, enmienda v1.3. Antes era texto libre.' }
                codigo_barras: { type: string, nullable: true }
                unidad_venta_codigo: { type: string, description: 'FK a unidad_medida — RF-CVI-025, enmienda v1.3' }
                unidad_inventario_codigo: { type: string, description: 'FK a unidad_medida — RF-CVI-025, enmienda v1.3' }
                es_fraccionable: { type: boolean }
                factor_conversion: { type: number, nullable: true }
                es_perecedero: { type: boolean, description: 'opcional desde la enmienda v1.3 — si se omite, se precarga desde categoria.es_perecedero' }
      responses:
        '201': { description: Producto creado }
        '422': { description: 'es_fraccionable=true sin factor_conversion (RN-CVI-001), o es_fraccionable=true con una unidad_venta que no admite decimales (RN-CVI-006, enmienda v1.3)' }
        '404': { description: 'categoria_id o alguno de los códigos de unidad no existe en su catálogo (enmienda v1.3)' }
  /productos/{id}:
    patch:
      summary: Actualizar producto (RF-CVI-007)
      operationId: actualizarProducto
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, properties: { nombre: { type: string }, categoria_id: { type: integer }, unidad_venta_codigo: { type: string }, unidad_inventario_codigo: { type: string }, es_fraccionable: { type: boolean }, factor_conversion: { type: number, nullable: true }, es_perecedero: { type: boolean }, activo: { type: boolean } } }
      responses:
        '200': { description: Actualizado }
        '422': { description: 'es_fraccionable=true sin factor_conversion (RN-CVI-001), o unidad_venta sin permite_decimales (RN-CVI-006)' }
        '404': { description: 'categoria_id o código de unidad inexistente (enmienda v1.3)' }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
...